
\name{blkdiag}
\alias{blkdiag}

\title{Block diagonal concatenation of matrix}

\usage{blkdiag(x)}

\arguments{
\item{x}{a list of matrix}
}
\value{
E - Block diagonal concatenation of matrix
}
\description{

  Y = BLKDIAG(A,B,...)  produces  diag(A,B,...)

}
\author{
  Yusen Zhang \cr
    \email{yusenzhang@126.com}
    }
