\name{mfeat_pix}
\alias{mfeat_pix}
\docType{data}
\title{
mfeat_pix dataset
}
\description{
This dataset consists of features of handwritten numerals (`0'--`9') extracted from a collection of Dutch utility maps. 200 patterns per class (for a total of 2,000 patterns) have been digitized in binary images. This dataset is about 240 pixel averages in 2 x 3 windows
}
\usage{data("mfeat_pix")}
\format{
  A data frame with 2000 observations on the following 240 variables.
}

\examples{
data(mfeat_pix)
}

\source{
https://archive.ics.uci.edu/ml/datasets/Multiple+Features
}
\keyword{datasets}
